Remove unused files & move WaveParser to wav package
