--- lib/aiff/index.ts
@@ -1,3 +1,0 @@
-import {AIFFParser} from "./AiffParser";
-
-export default AIFFParser;
