--- lib/apev2/index.ts
@@ -1,3 +1,0 @@
-import {APEv2Parser} from "./APEv2Parser";
-
-export default APEv2Parser;
