--- lib/asf/index.ts
@@ -1,3 +1,0 @@
-import {AsfParser} from "./AsfParser";
-
-export default AsfParser;
