--- lib/flac/index.ts
@@ -1,3 +1,0 @@
-import {FlacParser} from "./FlacParser";
-
-export default FlacParser;
