--- lib/mp4/index.ts
@@ -1,3 +1,0 @@
-import {MP4Parser} from "./MP4Parser";
-
-export default MP4Parser;
