--- lib/mpeg/index.ts
@@ -1,3 +1,0 @@
-import {MpegParser} from "./MpegParser";
-
-export default MpegParser;
