--- lib/ogg/index.ts
@@ -1,3 +1,0 @@
-import {OggParser} from './OggParser';
-
-export default OggParser;
