--- lib/riff/index.ts
@@ -1,3 +1,0 @@
-import { WaveParser } from './WaveParser';
-
-export default WaveParser;
