+++ lib/wav/WaveParser.ts
@@ -2,7 +2,7 @@ import * as strtok3 from 'strtok3/lib/core';
 import * as Token from 'token-types';
 import * as initDebug from 'debug';
 
-import * as riff from './RiffChunk';
+import * as riff from '../riff/RiffChunk';
 import * as WaveChunk from './../wav/WaveChunk';
 import { ID3v2Parser } from '../id3v2/ID3v2Parser';
