--- lib/wavpack/index.ts
@@ -1,3 +1,0 @@
-import {WavPackParser} from "./WavPackParser";
-
-export default WavPackParser;
